+++ MultiSubDownloader/src/main/java/org/lodder/subtools/multisubdownloader/lib/Info.java
@@ -1,7 +1,7 @@
 package org.lodder.subtools.multisubdownloader.lib;
 
 import org.lodder.subtools.multisubdownloader.settings.model.Settings;
-import org.lodder.subtools.sublibrary.model.Subtitle.SubtitleSource;
+import org.lodder.subtools.sublibrary.model.SubtitleSource;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
