+++ MultiSubDownloader/src/main/java/org/lodder/subtools/multisubdownloader/lib/control/subtitles/sorting/replacers/GroupReplacer.java
@@ -1,6 +1,7 @@
 package org.lodder.subtools.multisubdownloader.lib.control.subtitles.sorting.replacers;
 
 import java.util.Map;
+
 import org.lodder.subtools.sublibrary.model.Release;
 
 public class GroupReplacer implements KeywordReplacer {
