+++ MultiSubDownloader/src/main/java/org/lodder/subtools/multisubdownloader/subtitleproviders/opensubtitles/api/v2/OpenSubtitlesExecuter.java
@@ -1,6 +1,7 @@
 package org.lodder.subtools.multisubdownloader.subtitleproviders.opensubtitles.api.v2;
 
 import org.opensubtitles.invoker.ApiException;
+
 import com.pivovarit.function.ThrowingSupplier;
 
 public abstract class OpenSubtitlesExecuter {
