+++ MultiSubDownloader/src/test/java/org/lodder/subtools/multisubdownloader/lib/control/subtitles/sorting/SortWeightTest.java
@@ -1,14 +1,14 @@
 package org.lodder.subtools.multisubdownloader.lib.control.subtitles.sorting;
 
+import static org.junit.Assert.*;
+import static org.mockito.Mockito.*;
+
 import java.util.HashMap;
 import java.util.Map;
+
 import org.junit.Before;
 import org.junit.Test;
 import org.lodder.subtools.sublibrary.model.Release;
-
-import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.when;
-import static org.junit.Assert.*;
 
 public class SortWeightTest {
