+++ SubLibrary/src/test/java/org/lodder/subtools/sublibrary/control/RomanTest.java
@@ -1,6 +1,6 @@
 package org.lodder.subtools.sublibrary.control;
 
-import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.*;
 
 import org.junit.Before;
 import org.junit.Test;
